a1_jump_optimization/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
casadi>=3.6.0
matplotlib>=3.4.0
pandas>=1.3.0
pyyaml>=5.4.0
tqdm>=4.62.0
pytest>=7.0.0
pillow>=8.3.0
